rearranged some files for better modularity; started on modules list
